refactor: components/common/inquiry -> components/inquiry/list 로 구조 개선
