+++ src/components/inquiry/list/TeamTabs.tsx
